fix: bug when using single section in ebay_rest.json and not specifying section name (attempt 2) (#64)
In Py3, dict_keys objects are iterable but not indexable; in the single-value case convert to a tuple before retrieving 'first' (only) value.

Trying to create an API object without specifying the sections in ebay_rest.json (permitted when there is only one section in the ebay_rest.json file for applications/user etc) fails. This fixes that issue.

Secondly, 'applications' is hard-coded in which means

Arguably there should be a check that the single 'application' section has the same name as the single 'user' section etc., but I have not implemented this. If you just left the `sections = config_contents['applications'].keys()` line as it was, this would be enforced, but it would fail with a KeyError if the sections weren't aimed for (an additional check would be needed). I went for simplicity and decided it probably didn't really matter...